点赞、商家、评论、用户接口修改

+++ live-service/src/main/java/com/xq/live/dao/UserMapper.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface UserMapper{
     int deleteByPrimaryKey(Long id);
 
-    Long insert(User record);
+    int insert(User record);
 
     int insertSelective(User record);
